Masterpassword Security
Now the Program will only give the right password if the Masterpassword is provided. And the program will
check wheter it is the right M.P. if not it'll not even print anything. Only as feature. 
Due to this even an attacker that has access to the pass.json has can't get your password